Code refactoring
Split code into packages
