--- cmbert/multimodal_dataloader.py
@@ -1,19 +1,0 @@
-import sys
-import os
-import re
-import numpy as np
-
-
-
-
-# import mmsdk
-
-
-
-
-class MultimodalDataLoader():
-
-    def __init__(self, config):
-        super(MultimodalDataLoader, self).__init__()
-
-
