--- cmbert/utils.py
@@ -1,28 +1,0 @@
-import numpy as np
-
-def avg(intervals: np.array, features: np.array) -> np.array:
-    try:
-        return np.average(features, axis=0)
-    except:
-        return features
-    
-
-def bword_vector_2_sentence(bwords, filter = [b'sp'], encoding='utf-8'):
-    return ' '.join([bw.decode(encoding) for bw in bwords if bw not in filter])
-
-
-def group_texts(examples, chunk_size):
-    # Concatenate all texts
-    concatenated_examples = {k: sum(examples[k], []) for k in examples.keys()}
-    # Compute length of concatenated texts
-    total_length = len(concatenated_examples[list(examples.keys())[0]])
-    # We drop the last chunk if it's smaller than chunk_size
-    total_length = (total_length // chunk_size) * chunk_size
-    # Split by chunks of max_len
-    result = {
-        k: [t[i : i + chunk_size] for i in range(0, total_length, chunk_size)]
-        for k, t in concatenated_examples.items()
-    }
-    # Create a new labels column
-    result["labels"] = result["input_ids"].copy()
-    return result
