+++ mmanalysis/utils/__init__.py
@@ -1,0 +1,9 @@
+from .utils import (
+    avg,
+    bword_vector_2_sentence,
+    group_texts,
+    set_optimizer_custom_parameters,
+    get_criterion,
+    get_optimizer,
+    save_result,
+)
