+++ train_class.py
@@ -1,20 +1,37 @@
-from collections import defaultdict
 import copy 
-import json
 
-import numpy as np
 import torch
-from torch import nn
 from torch.utils.data import DataLoader
-from torch.optim import AdamW, SGD, Adam, RMSprop, Adadelta
 import transformers
-from transformers import get_scheduler, AutoTokenizer, DataCollatorWithPadding, DefaultDataCollator
+from transformers import get_scheduler
 from tqdm.auto import tqdm
 import evaluate
-from transformers import AutoModel, AutoTokenizer, BertModel, PreTrainedModel, BertConfig, BertPreTrainedModel
 
-import cmbert
-from mmsdk.mmdatasdk.dataset.standard_datasets.CMU_MOSI import cmu_mosi_std_folds as standard_folds
+from mmanalysis.datacollators import MMSeqClassDataCollator
+
+from mmanalysis.datasets.cmu import (
+    CmuDataset,
+    standard_folds
+)
+
+# from mmsdk.mmdatasdk.dataset.standard_datasets.CMU_MOSI import cmu_mosi_std_folds as standard_folds
+from mmanalysis.models.cmbert import (
+    CMBertForSequenceClassification,
+    CMBertTokenizer,
+)
+
+from mmanalysis.models.mmbert import (
+    MMBertForSequenceClassification,
+    MMBertTokenizer,
+)
+
+from mmanalysis.utils import (
+    set_optimizer_custom_parameters,
+    get_criterion,
+    get_optimizer,
+    save_result,
+)
+
 
 class TrainingArgs():
 
@@ -43,31 +60,6 @@ class TrainingArgs():
         self.best_model_path = best_model_path
         self.save_model_dest = save_model_dest
 
-    def __str__(self) -> str:
-        return str(self.__dict__.items())
-
-class MultiModalDataCollator():
-
-    def __init__(self, checkpoint, num_labels=2 ,device=torch.device("cpu")) -> None:
-        self.tokenizer = cmbert.MultimodalTokenizer(checkpoint=checkpoint)
-        self.device = device
-        if num_labels == 1:
-            self.labels_dtype = torch.float32
-        else:
-            self.labels_dtype = torch.long
-
-    def __call__(self, *args: torch.Any, **kwds: torch.Any) -> torch.Any:
-        rows = args[0]
-        batch = {key: [row[key] for row in rows] for key in rows[0]}
-        if 'audio_feat' not in batch:
-            batch['audio_feat'] = None
-        if 'visual_feat' not in batch:
-            batch['visual_feat'] = None
-
-        tokenized = self.tokenizer(text=batch['text_feat'], audio=batch['audio_feat'], visual=batch['visual_feat'], padding=True, truncation=True)
-        tokenized['labels'] = torch.tensor(batch['labels'], dtype=self.labels_dtype)
-        tokenized.to(self.device)
-        return  tokenized
 
 def prepare_data_splits(ds, num_labels):
     ds.labels_2_class(num_classes=num_labels)
@@ -83,15 +75,12 @@ def prepare_data_splits(ds, num_labels):
     ds.computational_sequences_2_array()
     ds.words_2_sentences()
 
-    train_ds = cmbert.CmuDataset.from_dataset(ds, fold='train')
-    test_ds = cmbert.CmuDataset.from_dataset(ds, fold='test')
-    valid_ds = cmbert.CmuDataset.from_dataset(ds, fold='valid')
+    train_ds = CmuDataset.from_dataset(ds, fold='train')
+    test_ds = CmuDataset.from_dataset(ds, fold='test')
+    valid_ds = CmuDataset.from_dataset(ds, fold='valid')
         
     return train_ds, valid_ds, test_ds
 
-
-from sklearn.metrics import mean_absolute_error as MAE
-from sklearn.metrics import mean_squared_error as MSE
 
 def evaluation(model, dataloader, criterion, metrics_names, task):
     model.eval()
@@ -175,30 +164,6 @@ def train(model, train_dataloader, valid_dataloader, num_epochs, optimizer, lr_s
 
     return best_model, train_loss, valid_eval
 
-def get_criterion(criterion_name):
-    if criterion_name == 'crossentropyloss':
-        return nn.CrossEntropyLoss()
-    elif criterion_name == 'mseloss': # ...
-        return nn.MSELoss()
-    elif criterion_name == 'maeloss':
-        return nn.L1Loss()
-    
-
-def get_optimizer(optimizer_name, **kwargs):
-    if optimizer_name == 'adamw':
-        return AdamW(**kwargs)
-    elif optimizer_name == 'adam':
-        return Adam(**kwargs)
-    elif optimizer_name == 'rmsprop':
-        return RMSprop(**kwargs)
-    elif optimizer_name == 'adadelta':
-        return Adadelta(**kwargs)
-
-def save_result(result, path):
-    with open(path, 'a+') as fd:
-        json.dump(result, fd)
-        fd.write('\n')
-        
 
 def main(dataset_config, model_config, training_arguments, results_path='experiments/results.jsonl', dsdeploy=False):
     device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
@@ -210,12 +175,13 @@ def main(dataset_config, model_config, training_arguments, results_path='experim
     
 
     # dataset
-    ds = cmbert.CmuDataset(dataset_config)
+    ds = CmuDataset(dataset_config)
     if dsdeploy:
         ds.deploy()
     
     # model
-    model = cmbert.CMBertForSequenceClassification(config=model_config)
+    model = CMBertForSequenceClassification(config=model_config)
+    # model = MMBertForSequenceClassification(config=model_config)
     if model_config.num_classes == 1:
         task = 'r'
     elif model_config.num_classes == 2:
@@ -229,20 +195,13 @@ def main(dataset_config, model_config, training_arguments, results_path='experim
     batch_size = training_arguments.batch_size
     criterion = get_criterion(training_arguments.criterion)
 
-    # Prepare optimizer
+
+    # Prepare optimizer parameters
     if ('audio_feat' in dataset_config.feature_names 
         or 'visual_feat' in dataset_config.feature_names):
-        param_optimizer = list(model.named_parameters())
-        no_decay = ['bias', 'LayerNorm.bias', 'LayerNorm.weight', 'norm.bias', 'norm.weight']
-        new_decay = ['modality_fusion']
-
-        optimizer_grouped_parameters = [
-            {'params': [p for n, p in param_optimizer if not any(nd in n for nd in no_decay) and not any(np in n for np in new_decay)], 'weight_decay': 0.01},
-            {'params': [p for n, p in param_optimizer if any(nd in n for nd in no_decay)], 'weight_decay': 0.0},
-            {'params':[p for n, p in param_optimizer if not any(nd in n for nd in no_decay )and any(np in n for np in new_decay)],'lr':0.01}
-        ]
-
-        parameters = optimizer_grouped_parameters
+        parameters = set_optimizer_custom_parameters(
+            model=model,
+        )
     else:
         parameters = model.parameters()
 
@@ -256,7 +215,10 @@ def main(dataset_config, model_config, training_arguments, results_path='experim
     model.to(device)
     criterion.to(device)
 
-    data_collator = MultiModalDataCollator(checkpoint=model_config.encoder_checkpoint, num_labels=model_config.num_classes, device=device)
+    # tokenizer = MMBertTokenizer(checkpoint=model_config.encoder_checkpoint)
+    tokenizer = CMBertTokenizer(checkpoint=model_config.encoder_checkpoint)
+
+    data_collator = MMSeqClassDataCollator(tokenizer=tokenizer, num_labels=model_config.num_classes, device=device)
     train_dataloader = DataLoader(train_ds, shuffle=True, batch_size=batch_size, collate_fn=data_collator)
     valid_dataloader = DataLoader(valid_ds, shuffle=True, batch_size=batch_size, collate_fn=data_collator)
     test_dataloader = DataLoader(test_ds, batch_size=batch_size, collate_fn=data_collator)
@@ -290,7 +252,6 @@ def main(dataset_config, model_config, training_arguments, results_path='experim
         metrics=metrics,
         task=task,
         best_model_metric=model_config.best_model_metric,
-        # save_best_model=training_arguments.save_best_model,
     )
 
     best_model = model
@@ -323,7 +284,3 @@ def main(dataset_config, model_config, training_arguments, results_path='experim
     }
 
     save_result(result, results_path)
-
-    # print(f'text_weight_1: {best_model.modality_fusion.text_weight_1}')
-    # print(f'audio_weight_1: {best_model.modality_fusion.audio_weight_1}')
-    # print(f'visual_weight_1: {best_model.modality_fusion.visual_weight_1}')
