+++ train_mlm.py
@@ -1,34 +1,49 @@
-from collections import defaultdict
+import numpy as np
 import copy 
-import json
 import math
 from datetime import datetime
 
 import numpy as np
 import torch
-from torch import nn
 from torch.utils.data import DataLoader
-from torch.optim import AdamW, SGD, Adam, RMSprop, Adadelta
 import transformers
-from transformers import get_scheduler, AutoTokenizer, DataCollatorWithPadding, DefaultDataCollator
+from transformers import get_scheduler
 from tqdm.auto import tqdm
-import evaluate
-from transformers import AutoModel, AutoTokenizer, BertModel, PreTrainedModel, BertConfig, BertPreTrainedModel, AutoModelForMaskedLM
+from datasets.dataset_dict import DatasetDict
+from datasets import Dataset
 
 import model_config
-import cmbert.masked_language_modeling as mlm
-
-import cmbert
-from mmsdk.mmdatasdk.dataset.standard_datasets.CMU_MOSI import cmu_mosi_std_folds as standard_folds
-from cmbert import SDK_DS
 
 
-import collections
-import numpy as np
+from mmanalysis.models.cmbert import (
+    CMBertConfig,
+    CMBertForMaskedLM,
+    CMBertTokenizer,
+)
+from mmanalysis.models.mmbert import (
+    MMBertConfig,
+    MMBertForMaskedLM,
+    MMBertTokenizer,
+)
 
-from transformers import default_data_collator
+from mmanalysis.datacollators import MMMLMDataCollator
 
-wwm_probability = 0.2
+from mmanalysis.datasets.cmu import (
+    CmuDataset,
+    standard_folds,
+    SDK_DS,
+    CmuDatasetConfig
+)
+
+from mmanalysis.utils import (
+    get_criterion,
+    get_optimizer,
+    group_texts,
+    save_result,
+    set_optimizer_custom_parameters
+)
+
+# from transformers import default_data_collator
 
 class TrainingArgs():
 
@@ -61,105 +76,25 @@ class TrainingArgs():
         return str(self.__dict__.items())
 
 
-class MLMDataCollator():
+def prepare_data_splits(ds):
 
-    def __init__(self, checkpoint, return_tensors='list', device=torch.device('cpu')) -> None:
-        self.tokenizer = cmbert.MultimodalTokenizer(checkpoint=checkpoint)
-        self.device = device
-        self.return_tensors = return_tensors
+    folds = {
+        'train': standard_folds.standard_train_fold,
+        'valid': standard_folds.standard_valid_fold,
+        'test': standard_folds.standard_test_fold,
+        }
+   
+    ds.train_test_valid_split(folds=folds)
+    ds.replace_inf_and_nan_values(value=0)
+    ds.computational_sequences_2_array()
+    ds.words_2_sentences()
 
-    def __call__(self, *args: torch.Any, **kwds: torch.Any) -> torch.Any:
-        rows = args[0]
-        batch = {key: [row[key] for row in rows] for key in rows[0]}
+    train_ds = CmuDataset.from_dataset(ds, fold='train')
+    test_ds = CmuDataset.from_dataset(ds, fold='test')
+    valid_ds = CmuDataset.from_dataset(ds, fold='valid')
+        
+    return train_ds, valid_ds, test_ds
 
-
-        features = []
-        for i in range(len(batch['input_ids'])):
-            features.append({})
-            for k, v in batch.items():
-                features[i][k] = v[i]     
-
-        batch = self.whole_word_masking_data_collator(batch)
-    
-        return batch
-
-
-    def whole_word_masking_data_collator(self, batch):
-        tokenizer = self.tokenizer
-
-        word_ids_list = batch.pop('word_ids')
-        labels_list = batch['labels']
-
-        for i in range(len(batch['input_ids'])):
-            word_ids = word_ids_list[i]
-            labels = labels_list[i]
-
-            mapping = collections.defaultdict(list)
-            current_word_index = -1
-            current_word = None
-            for idx, word_id in enumerate(word_ids):
-                if word_id is not None:
-                    if word_id != current_word:
-                        current_word = word_id
-                        current_word_index += 1
-                    mapping[current_word_index].append(idx)
-
-            # Randomly mask words
-            np.random.seed(7)
-            mask = np.random.binomial(1, wwm_probability, (len(mapping),))
-            new_labels = [-100] * len(labels)
-            for word_id in np.where(mask)[0]:
-                word_id = word_id.item()
-                for idx in mapping[word_id]:
-                    new_labels[idx] = labels[idx]
-                    batch['input_ids'][i][idx] = tokenizer.mask_token_id
-            batch["labels"][i] = new_labels
-
-        for key, value in batch.items():
-            if self.return_tensors == 'pt':
-                if key == 'labels':
-                    batch[key] = torch.tensor(value, dtype=torch.long).to(self.device)
-                else:
-                    batch[key] = torch.tensor(value).to(self.device)
-            elif self.return_tensors == 'np':
-                batch[key] = np.array(value)
-
-        return batch
-     
-
-    def whole_word_masking_data_collator_bck(self, features):
-        tokenizer = self.tokenizer
-        for feature in features:
-            for key in feature:
-                feature[key] = np.array(feature[key])
-
-            word_ids = feature.pop("word_ids")
-
-            # Create a map between words and corresponding token indices
-            mapping = collections.defaultdict(list)
-            current_word_index = -1
-            current_word = None
-            for idx, word_id in enumerate(word_ids):
-                if word_id is not None:
-                    if word_id != current_word:
-                        current_word = word_id
-                        current_word_index += 1
-                    mapping[current_word_index].append(idx)
-
-            # Randomly mask words
-            np.random.seed(7)
-            mask = np.random.binomial(1, wwm_probability, (len(mapping),))
-            input_ids = feature["input_ids"]
-            labels = feature["labels"]
-            new_labels = [-100] * len(labels)
-            for word_id in np.where(mask)[0]:
-                word_id = word_id.item()
-                for idx in mapping[word_id]:
-                    new_labels[idx] = labels[idx]
-                    input_ids[idx] = tokenizer.mask_token_id
-            feature["labels"] = new_labels
-
-        return features
 
 def tokenize_function(batch, *args, **kwargs):
     audio = [np.array(element) for element in batch['audio_feat']]
@@ -171,30 +106,6 @@ def tokenize_function(batch, *args, **kwargs):
         tokenized["word_ids"] = [np.array(tokenized.word_ids(i)) for i in range(len(tokenized["input_ids"]))]
 
     return tokenized
-
-def group_texts(examples):
-    chunk_size = 128
-
-    # # Concatenate all texts
-    concatenated_examples = {k: sum(examples[k], []) for k in examples.keys()}
-    # Compute length of concatenated texts
-    total_length = len(concatenated_examples['input_ids'])
-    # We drop the last chunk if it's smaller than chunk_size
-    total_length = (total_length // chunk_size) * chunk_size
-    # Split by chunks of max_len
-    result = {
-        k: [t[i : i + chunk_size] for i in range(0, total_length, chunk_size)]
-        for k, t in concatenated_examples.items()
-    }
-    # Create a new labels column
-    result["labels"] = result["input_ids"].copy()
-
-    return result
-
-def save_result(result, path):
-    with open(path, 'a+') as fd:
-        json.dump(result, fd)
-        fd.write('\n')
 
 def evaluation(model, dataloader, metrics_names=None):
     model.eval()
@@ -218,7 +129,6 @@ def evaluation(model, dataloader, metrics_names=None):
 
 # def train(num_epochs,):
 def train(model, train_dataloader, valid_dataloader, num_epochs, optimizer, lr_scheduler, criterion=None, metrics=None, task=None, best_model_metric='perplexity'):
-    pass
     num_training_steps = num_epochs * len(train_dataloader)
     progress_bar = tqdm(range(num_training_steps))
     train_loss = []
@@ -268,23 +178,26 @@ def main(dataset_config, _model_config, training_arguments, results_path='experi
     if device == torch.device('cuda'):
         torch.cuda.manual_seed_all(seed)
 
-    tokenizer = cmbert.MultimodalTokenizer(checkpoint=model_config.encoder_checkpoint)
+    tokenizer = CMBertTokenizer(checkpoint=model_config.encoder_checkpoint)
 
     # dataset
-    ds = cmbert.CmuDataset(dataset_config)
+    ds = CmuDataset(dataset_config)
     if dsdeploy:
         ds.deploy()
 
     # model
 
     # split data into train, valid, test datasets
-    train_ds, valid_ds, test_ds = mlm.prepare_data_splits(ds=ds, num_labels=_model_config.num_classes)
+    train_ds, valid_ds, test_ds = prepare_data_splits(ds=ds)
 
-    dsdict = mlm.cmudataset_to_datasetdict(
-        train_ds=train_ds,
-        valid_ds=valid_ds,
-        test_ds=test_ds,
-        )
+    dsdict = DatasetDict({
+        'train': Dataset.from_dict(train_ds.dataset),
+        'valid': Dataset.from_dict(valid_ds.dataset),
+        'test': Dataset.from_dict(test_ds.dataset),
+        # 'train': Dataset.from_dict(train_ds[:225]),
+        # 'valid': Dataset.from_dict(valid_ds[:225]),
+        # 'test': Dataset.from_dict(test_ds[:225]),
+    })
 
     tokenized_datasets = dsdict.map(
         tokenize_function, 
@@ -294,25 +207,39 @@ def main(dataset_config, _model_config, training_arguments, results_path='experi
     )
 
     lm_datasets = tokenized_datasets.map(
-        group_texts, 
-        batched=True
+        group_texts,
+        batched=True,
+        # fn_kwargs={'chunk_size': training_arguments.chunk_size}
+        fn_kwargs={'chunk_size': 128}
     )
     
     device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
-    data_collator = MLMDataCollator(_model_config.encoder_checkpoint, return_tensors='pt', device=device)
-    # data_collator = MLMDataCollator(_model_config.encoder_checkpoint, return_tensors='list')
-
+    data_collator = MMMLMDataCollator(tokenizer=tokenizer, return_tensors='pt', device=device)
+    
     batch_size = 8
   
     train_dataloader = DataLoader(lm_datasets['train'], shuffle=True, batch_size=batch_size, collate_fn=data_collator)
     valid_dataloader = DataLoader(lm_datasets['valid'], shuffle=True, batch_size=batch_size, collate_fn=data_collator)
     test_dataloader = DataLoader(lm_datasets['test'], batch_size=batch_size, collate_fn=data_collator)
 
-    model = cmbert.CMBertForMaskedLM(_model_config)
+    model = CMBertForMaskedLM(_model_config)
     model.to(device)
 
-    optimizer = AdamW(model.parameters(), lr=5e-5)
-    # num_epochs = 3
+    # Prepare optimizer
+    if ('audio_feat' in dataset_config.feature_names 
+        or 'visual_feat' in dataset_config.feature_names):
+        parameters = set_optimizer_custom_parameters(
+            model=model,
+        )
+    else:
+        parameters = model.parameters()
+
+    optimizer = get_optimizer(
+        optimizer_name=training_arguments.optimizer,
+        params=parameters,
+        lr=training_arguments.lr
+    )
+
     num_update_steps_per_epoch = len(train_dataloader)
     num_training_steps = training_arguments.num_epochs * num_update_steps_per_epoch
 
@@ -356,10 +283,6 @@ def main(dataset_config, _model_config, training_arguments, results_path='experi
             state_dict=best_model.state_dict(),
         )
 
-    with open('model_params.txt', 'w+') as fd:
-        for param in best_model.parameters():
-            fd.write(str(param))
-
 if __name__ =='__main__':
     ds = 'CMUMOSI'
     text_features = SDK_DS[ds]['RAWTEXT']['featuresname']
@@ -367,7 +290,7 @@ if __name__ =='__main__':
     visual_features = SDK_DS[ds]['FACET42']['featuresname']
     labels = SDK_DS[ds]['LABELS']['featuresname']
 
-    dataset_config = cmbert.CmuDatasetConfig(
+    dataset_config = CmuDatasetConfig(
         sdkpath = r'D:\Studia\magisterskie\Praca_magisterska\data\repo\CMU-MultimodalSDK',
         dataset = ds,
         text_features = text_features,
@@ -378,7 +301,7 @@ if __name__ =='__main__':
         load_preprocessed = True,
     )
 
-    _model_config = cmbert.CMBertConfig(
+    _model_config = CMBertConfig(
         encoder_checkpoint=model_config.encoder_checkpoint,
         modality_att_dropout_prob=model_config.modality_att_dropout_prob,
         hidden_dropout_prob=model_config.hidden_dropout_prob,
@@ -406,15 +329,3 @@ if __name__ =='__main__':
 
     main(dataset_config=dataset_config, _model_config=_model_config, training_arguments=training_arguments, results_path=results_path)
     
-    # device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
-
-    # path = 'models/distilbert-base-uncased_05May2024_215644'
-    # best_model = cmbert.CMBertForMaskedLM.from_pretrained(path).to(device)
-
-
-    # model2 = cmbert.CMBertForSequenceClassification.from_pretrained(path).to(device)
-
-    
-    # with open('model_params3.txt', 'w+') as fd:
-    #     for param in model2.parameters():
-    #         fd.write(str(param))
